requirements.txt
google-generativeai>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
schedule>=1.2.0
playwright>=1.40.0
